+++ cnccode/business_classes/BUActivity.inc.php
@@ -11120,6 +11120,20 @@ is currently a balance of ';
         return $dbejCallactivity;
     }
 
+    public function getProblemsByContact($contactID)
+    {
+        $dbeJProblem = new DBEJProblem($this);
+
+        $dbeJProblem->getProblemsByContactID($contactID);
+
+        $this->getData(
+            $dbeJProblem,
+            $dsResults
+        );
+
+        return $dsResults;
+    }
+
 
 } // End of class
 ?>