+++ cnccode/controller_classes/CTActivity.inc.php
@@ -261,6 +261,9 @@ class CTActivity extends CTCNC
                     }
                 }
                 break;
+            case 'displayServiceRequestForContactPopup':
+                $this->serviceRequestsForContactPopup();
+                break;
             case CTACTIVITY_ACT_DISPLAY_ACTIVITY:
                 $this->displayActivity();
                 break;
@@ -2147,6 +2150,9 @@ class CTActivity extends CTCNC
                 'contractListPopupLink'              => $this->getContractListPopupLink(
                     $dsCallActivity->getValue('customerID')
                 ),
+                'contactHistoryLink'                 => $this->getServiceRequestForContactLink(
+                    $dsCallActivity->getValue(DBECallActivity::contactID)
+                ),
                 'salesOrderLink'                     => $this->getSalesOrderLink(
                     $dsCallActivity->getValue('linkedSalesOrderID')
                 ),
@@ -3059,7 +3065,8 @@ class CTActivity extends CTCNC
 
         $_SESSION[$this->sessionKey]['internalNotes'] = $_REQUEST['internalNotes'];
 
-        /* Loop here */
+
+        $dsContactSrs = $this->buActivity->getProblemsByContact($_REQUEST['contactID']);
 
         $dsActiveSrs = $this->buActivity->getActiveProblemsByCustomer($_REQUEST['customerID']);
 
@@ -3068,11 +3075,78 @@ class CTActivity extends CTCNC
             'ActivityExistingRequests.inc'
         );
 
+        $dbeContact = new DBEContact($this);
+        $dbeContact->getRow($_REQUEST['contactID']);
+
+
         $this->template->set_var(
             [
-                'techNotes' => $dsCustomer->getValue(DBECustomer::techNotes)
+                'techNotes'   => $dsCustomer->getValue(DBECustomer::techNotes),
+                'contactName' => $dbeContact->getValue(DBEContact::firstName) . " " . $dbeContact->getValue(
+                        DBEContact::lastName
+                    )
             ]
         );
+
+        $this->template->set_block(
+            'ActivityExistingRequests',
+            'contactProblemBlock',
+            'contactProblems'
+        );
+
+        while ($dsContactSrs->fetchNext()) {
+
+            $urlCreateFollowOn =
+                $this->buildLink(
+                    'Activity.php',
+                    array(
+                        'action'         => 'createFollowOnActivity',
+                        'callActivityID' => $dsContactSrs->getValue('lastCallActivityID'),
+                        'reason'         => $_REQUEST['reason']
+                    )
+                );
+
+            $urlProblemHistoryPopup =
+                $this->buildLink(
+                    'Activity.php',
+                    array(
+                        'action'    => 'problemHistoryPopup',
+                        'problemID' => $dsContactSrs->getValue('problemID'),
+                        'htmlFmt'   => CT_HTML_FMT_POPUP
+                    )
+                );
+
+            $this->template->set_var(
+                array(
+                    'contactProblemID'              => $dsContactSrs->getValue("problemID"),
+                    'contactDateRaised'             => Controller::dateYMDtoDMY($dsContactSrs->getValue('dateRaised')),
+                    'contactReason'                 => self::truncate(
+                        $dsContactSrs->getValue("reason"),
+                        100
+                    ),
+                    'contactLastReason'             => self::truncate(
+                        $dsContactSrs->getValue("lastReason"),
+                        100
+                    ),
+                    'contactEngineerName'           => $dsContactSrs->getValue("engineerName"),
+                    'createFollowOnLink'            => $dsContactSrs->getValue(
+                        DBEJProblem::status
+                    ) == 'C' ? '' : "<a href=" . $urlCreateFollowOn . ">Log activity</a>",
+                    'contactUrlProblemHistoryPopup' => $urlProblemHistoryPopup,
+                    'contactPriority'               => $dsContactSrs->getValue(DBEJProblem::priority),
+                    'contactPriorityClass'          => $dsContactSrs->getValue(
+                        DBEJProblem::priority
+                    ) == 1 ? 'class="redRow"' : null
+                )
+            );
+
+            $this->template->parse(
+                'contactProblems',
+                'contactProblemBlock',
+                true
+            );
+
+        }
 
         $this->template->set_block(
             'ActivityExistingRequests',
@@ -3762,6 +3836,92 @@ class CTActivity extends CTCNC
         exit;
     }    // end allocateAdditionalTime
 
+    function serviceRequestsForContactPopup()
+    {
+        $this->setTemplateFiles(
+            'ServiceRequestsForContactPopup',
+            'ServiceRequestsForContactPopup'
+        );
+        $dsContactSrs = $this->buActivity->getProblemsByContact($_REQUEST['contactID']);
+
+
+        $dbeContact = new DBEContact($this);
+        $dbeContact->getRow($_REQUEST['contactID']);
+        $this->setPageTitle(
+            'Service Requests For ' . $dbeContact->getValue(DBEContact::firstName) . " " . $dbeContact->getValue(
+                DBEContact::lastName
+            )
+        );
+
+        $this->template->set_block(
+            'ServiceRequestsForContactPopup',
+            'contactProblemBlock',
+            'contactProblems'
+        );
+
+        while ($dsContactSrs->fetchNext()) {
+
+            $urlCreateFollowOn =
+                $this->buildLink(
+                    'Activity.php',
+                    array(
+                        'action'         => 'createFollowOnActivity',
+                        'callActivityID' => $dsContactSrs->getValue('lastCallActivityID'),
+                        'reason'         => $_REQUEST['reason']
+                    )
+                );
+
+            $urlProblemHistoryPopup =
+                $this->buildLink(
+                    'Activity.php',
+                    array(
+                        'action'    => 'problemHistoryPopup',
+                        'problemID' => $dsContactSrs->getValue('problemID'),
+                        'htmlFmt'   => CT_HTML_FMT_POPUP
+                    )
+                );
+
+            $this->template->set_var(
+                array(
+                    'contactProblemID'              => $dsContactSrs->getValue("problemID"),
+                    'contactDateRaised'             => Controller::dateYMDtoDMY($dsContactSrs->getValue('dateRaised')),
+                    'contactReason'                 => self::truncate(
+                        $dsContactSrs->getValue("reason"),
+                        100
+                    ),
+                    'contactLastReason'             => self::truncate(
+                        $dsContactSrs->getValue("lastReason"),
+                        100
+                    ),
+                    'contactEngineerName'           => $dsContactSrs->getValue("engineerName"),
+                    'createFollowOnLink'            => $dsContactSrs->getValue(
+                        DBEJProblem::status
+                    ) == 'C' ? '' : "<a href=" . $urlCreateFollowOn . ">Log activity</a>",
+                    'contactUrlProblemHistoryPopup' => $urlProblemHistoryPopup,
+                    'contactPriority'               => $dsContactSrs->getValue(DBEJProblem::priority),
+                    'contactPriorityClass'          => $dsContactSrs->getValue(
+                        DBEJProblem::priority
+                    ) == 1 ? 'class="redRow"' : null
+                )
+            );
+
+            $this->template->parse(
+                'contactProblems',
+                'contactProblemBlock',
+                true
+            );
+
+        }
+        $this->template->parse(
+            'CONTENTS',
+            'ServiceRequestsForContactPopup',
+            true
+        );
+        $this->parsePage();
+
+        exit;
+    }    // end allocateAddition
+
     function customerProblemPopup()
     {
         $this->setTemplateFiles(
@@ -4128,6 +4288,9 @@ class CTActivity extends CTCNC
                 'passwordLink'                 => $this->getPasswordLink($dsCallActivity->getValue('customerID')),
                 'thirdPartyContactLink'        => $this->getThirdPartyContactLink(
                     $dsCallActivity->getValue('customerID')
+                ),
+                'contactHistoryLink'                 => $this->getServiceRequestForContactLink(
+                    $dsCallActivity->getValue(DBECallActivity::contactID)
                 ),
                 'generatePasswordLink'         => $this->getGeneratePasswordLink(),
                 'salesOrderLink'               => $this->getSalesOrderLink(
@@ -6812,6 +6975,20 @@ class CTActivity extends CTCNC
         $dbeContact->getRow($contactId);
 
         return $dbeContact->getValue(DBEContact::notes);
+    }
+
+    private function getServiceRequestForContactLink($contactID)
+    {
+        $contactHistory =
+            $this->buildLink(
+                'Activity.php',
+                array(
+                    'action'    => 'displayServiceRequestForContactPopup',
+                    'contactID' => $contactID,
+                    'htmlFmt'   => CT_HTML_FMT_POPUP
+                )
+            );
+        return '| <a href="#" title="Contact SR History" onclick="window.open(\'' . $contactHistory . '\', \'reason\', \'scrollbars=yes,resizable=yes,height=400,width=1225,copyhistory=no, menubar=0\')">Contact SR History</a>';
     }
 }